+++ packages/form-render/src/form-items/tree-select/index.tsx
@@ -95,7 +95,7 @@ export function renderTreeSelectItem<T=DataRecord>(options: RenderTreeSelectItem
       break
     }
     case isRef(options.options): {
-      watchOnce(options.options, () => {
+      watchOnce(options.options as Ref<SelectOptions>, () => {
         updateSelectOptions((options.options as Ref<SelectOptions>).value)
       }, {
         immediate: true,