+++ packages/form-render/src/index.ts
@@ -1,5 +1,6 @@
 import 'virtual:uno.css'
 import './style.less'
+import 'vue/jsx'
 
 export { FormRender, type FormRenderInstance } from './form-render'
 export * from './defines'