+++ packages/table-render/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gopowerteam/table-render
 
+## 0.0.109
+
+### Patch Changes
+
+- upgrade vue 3.3 to 3.4
+- Updated dependencies
+  - @gopowerteam/modal-render@0.0.40
+  - @gopowerteam/form-render@0.0.61
+
 ## 0.0.108
 
 ### Patch Changes